dom/src/threads/test/test_xhr.html

<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads XHR(Bug 450452 )
-->
<head>
  <title>Test for DOM Worker Threads XHR (Bug 450452 )</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=450452">DOM Worker Threads XHR (Bug 450452)</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("xhr_worker.js");

  worker.onmessage = function(event) {
    is(event.target, worker);
    var args = eval(event.data);
    switch (args.type) {
      case "progress": {
        ok(parseInt(args.current) <= parseInt(args.total));
      } break;
      case "error": {
        ok(false, "XHR error: " + args.error);
      } break;
      case "load": {
        is(args.data, "A noisy noise annoys an oyster.");
        document.getElementById("content").textContent = args.data;
      } break;
      case "loadend": {
        ok(true, "Should have got loadend.");
        SimpleTest.finish();
        break;
      }
      default: {
        ok(false, "Unexpected message");
        SimpleTest.finish();
      }
    }
  };

  worker.onerror = function(event) {
    is(event.target, worker);
    ok(false, "Worker had an error:" + event.data);
    SimpleTest.finish();
  }

  worker.postMessage("testXHR.txt");

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
